update rakefile example to run without erroring
This fixes an issue which would cause the Rake task to error out when running it with an app name other than `hello` and without first creating `packaging/` directory. These changes are to avoid user frustration when attempting to use this as a guide for packaging their own app, not just following the example.