Update Bundler from 1.9.9 -> 1.9.10 to prevent write permissions error
Bundler 1.9.10 contains a specific change to fix a conflict that occurs
when projects running Bundler 1.9.x and 1.10.x both execute `bundle
install`. With 1.10.x+, a `Gemfile.lock` contains a section that
specifies the Bundler version.

A Travelling Ruby package built using recent Bundler versions will
append a `BUNDLED WITH` section to their `Gemfile.lock`, but when the
executables are run with the suggested shim with Travelling Ruby,
`bundler/setup` will cause a write permissions error when Bundler 1.9.9
attempts to remove the `BUNDLED WITH` section from the end of the
Gemfile.

Relevant explanation by Bundler team:
http://bundler.io/blog/2015/06/24/version-1-10-released.html