Fix installing in path containing spaces
Installing to any directory with spaces in the path failed because of
GEM_HOME and GEM_PATH having not enough quotes around them and ruby
failing to parse the options from RUBYOPT when the -r option had spaces
in it.

Fixed this by adding the necessary escapes to add doublequotes around
the env variables and moved restore_environment.rb to the
lib/ruby/site_ruby/ folder, which is in the ruby load path, so the -r
option doesn't need to reference the full absolute path.

Fixes #38.